Random Enemy Movement 2
This code is important for the function of the health script. It also is slightly edited in the way that it changes the polar bear's direction and removes the collision timer in the previous version because it was not needed. The math that was needed for the general polar bear movement has been simplified and some unnecessary math has been removed.