health2
This is the second iteration of the health script. I completely reworked the health system so that the enemy movement script didn't contain any calculations for the health system, meaning I could use it for the other enemies, making it a universal script. I then transferred the code(with a few major, unnoticeable changes) to this script. Although a lot longer than the first health script, this second iteration works for any number of enemies chosen and is, therefore, better than the first one, which only functioned for one enemy.