difficulty carrier
This code is what collects the difficulty int from the difficulty selector script and then carries it over to the play scene,  where it is used to determine how many polar bears will spawn. The DontDestroyOnLoad(this.gameObject) is what allows this object (and therefore the difficulty selected) to be carried over the scene transfer.