spawn controller
This is actually part of another script created by my development partner, but this is one of the parts that I had to add in order for the polar bears to spawn. This section collects the difficulty from the game object that was carried over the scenes (see difficulty carrier) and uses that in value to spawn the correct number of polar bears.