Random Enemy Movement 3
This is essentially the same as the previous iteration, except that I have removed the integration of the health system, as it was not needed anymore. This is because I used another script to handle the distanceBewteen calculations etc because it was easier and allowed this script to be used over and over on different objects without any problems.