difficulty selector 2
This is the second iteration of the difficulty selector script. It is slightly less efficient but that is because this is the only way that it will work. The previous one allowed you to enter numbers higher than 3 and lower than 1, but it would just set the difficulty to a default value of 1. This script makes it so that if you enter any number apart from 1 2 or 3, it makes the play button disappear completely, meaning you have to enter 1 2 or 3 in order to be able to click the play button.